+++ src/main/java/be/ugent/idlab/knows/dataio/access/COMPRESSION.java
@@ -1,5 +1,0 @@
-package be.ugent.idlab.knows.dataio.access;
-
-public class COMPRESSION {
-    public static final String GZIP = "gzip";
-}