+++ src/main/java/be/ugent/idlab/knows/dataio/exceptions/HeaderEmptyValuesException.java
@@ -1,9 +1,12 @@
 package be.ugent.idlab.knows.dataio.exceptions;
 
+import java.io.Serial;
+
 /**
  * An exception to be thrown when the header contains empty values.
  */
 public class HeaderEmptyValuesException extends RuntimeException {
+    @Serial
     private static final long serialVersionUID = -5337665888088989438L;
 
     public HeaderEmptyValuesException(String path) {