+++ src/main/java/be/ugent/idlab/knows/dataio/streams/CSVWSourceStream.java
@@ -6,7 +6,7 @@ import be.ugent.idlab.knows.dataio.iterators.csvw.CSVWConfiguration;
 import be.ugent.idlab.knows.dataio.record.Record;
 
 import java.io.IOException;
-import java.sql.SQLException;
+import java.io.Serial;
 import java.util.Spliterator;
 import java.util.Spliterators;
 import java.util.stream.Stream;
@@ -17,6 +17,7 @@ import java.util.stream.StreamSupport;
  * Internally relies on CSVWSourceIterator for record generation.
  */
 public class CSVWSourceStream implements SourceStream {
+    @Serial
     private static final long serialVersionUID = -9054328656428580251L;
     private final CSVWSourceIterator iterator;