+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated rxjava to 3.1.6
 - Added source encoding in `pom.xml`
 - Removed unused dependency on jackson-annotations.
+- Removed dependency on commons-io
 
 ### Security
 - Updated OpenCSV to 5.8