+++ CHANGELOG.md
@@ -6,4 +6,35 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-Initial release.
+
+### Changed
+- Require Java 17 (or more recent)
+
+### Fixed
+- Updated Maven Surefire plugin to 3.1.2
+- Removed dependency on Logback
+- Updated POI to 5.2.3
+- Removed dependency on Commons CSV
+- Updated all testcontainers to 1.18.3
+- Updated mssql-jdbc to 12.2.1.jre11
+- Updated ojdbc11 to 23.2.0.0
+- Updated mybatis to 3.5.13
+- Updated jOpenDocument to 1.3
+- Updated xlsx-streamer to 2.2.0
+- Updated rxjava to 3.1.6
+- Added source encoding in `pom.xml`
+- Removed unused dependency on jackson-annotations.
+- Removed dependency on commons-io
+- Removed unused code
+- Use Jena to query local RDF files instead of Fuseki, eliminating need to start SPARQL endpoint.
+
+### Security
+- Updated OpenCSV to 5.8
+- Updated JSonPath to 2.8.0
+- Updated JSoup to 1.16.1
+- Updated PostgreSQL driver to 42.6.0
+- Updated mysql-connector to 8.1.0
+- Updated jsurfer-jackson to 1.6.4
+- Updated apache-any23-encoding to 2.7
+- Updated excel-streaming-reader to 4.0.5
+- Updated jena-fuseki-* to 4.9.0